Removed CanMigrate from MigrationTracker
CanMigrateTo does not make sens at MigrationTracker, because how should
it detect this? It can not call CanMigrate on all MigrationStep, because
each MigrationStep might depend on the other before. So each previous migration
step have to be applied, before the next one can be evaluated. This
contradicts the purpose of CanMigrate at MigrationTracker.